% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MountTom}
\alias{MountTom}
\title{MountTom Dinosaur Track Dataset}
\format{
A list consisting of two elements:
\itemize{
\item \strong{Trajectories}: A list of interpolated trajectories, where each trajectory is a series of midpoints between consecutive footprints.
\item \strong{Footprints}: A list of data frames containing footprint coordinates, metadata (e.g., image reference, ID), and a marker indicating whether the footprint is actual or inferred.
}
}
\source{
Ostrom, J. H. (1972). Were some dinosaurs gregarious?. Palaeogeography, Palaeoclimatology, Palaeoecology, 11(4), 287-301.
}
\usage{
MountTom
}
\description{
A 'track' R object representing dinosaur tracks from the Mount Tom site.
}
\keyword{datasets}
