% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuAnTeTrack-package.R
\docType{package}
\name{QuAnTeTrack-package}
\alias{QuAnTeTrack}
\alias{QuAnTeTrack-package}
\title{QuAnTeTrack: Quantitative Analysis of Tetrapod Trackways}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A quantitative and automated tool to extract (palaeo)biological information (i.e., measurements, velocities, similarity metrics, etc.) from the analysis of tetrapod trackways. Methods implemented in the package draw from several sources, including Alexander (1976) \doi{10.1038/261129a0}, Batschelet (1981, ISBN:9780120810505), Benhamou (2004) \doi{10.1016/j.jtbi.2004.03.016}, Bovet and Benhamou (1988) \doi{10.1016/S0022-5193(88)80038-9}, Cheung et al. (2007) \doi{10.1007/s00422-007-0158-0}, Cheung et al. (2008) \doi{10.1007/s00422-008-0251-z}, Cleasby et al. (2019) \doi{10.1007/s00265-019-2761-1}, Farlow et al. (1981) \doi{10.1038/294747a0}, Ostrom (1972) \doi{10.1016/0031-0182(72)90049-1}, Rohlf (2008) \url{https://sbmorphometrics.org/}, Rohlf (2009) \url{https://sbmorphometrics.org/}, Ruiz and Torices (2013) \doi{10.1080/10420940.2012.759115}, Scrucca et al. (2016) \doi{10.32614/RJ-2016-021}, Thulborn and Wade (1984) \url{https://www.museum.qld.gov.au/collections-and-research/memoirs/nature-21/mqm-n21-2-11-thulborn-wade}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MacroFunUV/QuAnTeTrack}
  \item \url{https://macrofunuv.github.io/QuAnTeTrack/}
  \item Report bugs at \url{https://github.com/MacroFunUV/QuAnTeTrack/issues}
}

}
\author{
\strong{Maintainer}: Humberto G Ferrón \email{humberto.ferron@uv.es} (\href{https://orcid.org/0000-0003-2254-8424}{ORCID}) [copyright holder]

}
\keyword{internal}
