% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_track.R
\name{subset_track}
\alias{subset_track}
\title{Subset trackways}
\usage{
subset_track(data, tracks = NULL)
}
\arguments{
\item{data}{A \code{trackway} R object, which is a list consisting of two elements:
\itemize{
\item \strong{\code{Trajectories}}: A list of trajectories representing trackway midlines,
interpolated by connecting the midpoints of successive left–right footprint pairs
(i.e., footprints linked by pace lines). Includes columns \code{X}, \code{Y},
\code{IMAGE}, \code{ID}, and \code{Side} (set to \code{"Medial"}).
\item \strong{\code{Footprints}}: A list of data frames containing footprint coordinates
and associated metadata, with a \code{Side} column (\code{"R"} or \code{"L"})
and a \code{missing} marker (\code{"Actual"} or \code{"Inferred"}).
}}

\item{tracks}{A numeric vector specifying the indices of trackways to subset. The default is to include all trackways.}
}
\value{
A \code{trackway} R object that contains only the specified subset of trackways. The structure of the returned object mirrors the input structure but includes only the selected trackways.
}
\description{
\code{subset_track()} is a function that subsets trackways from a \code{trackway} R object based on the specified indices.
}
\details{
This function subsets both the \strong{\code{Trajectories}} and \strong{\code{Footprints}} elements of the input data based on the provided vector of indices.
It allows users to focus on a specific subset of trackways for further analysis or visualization, particularly when working with large datasets containing numerous trackways.
}
\section{Logo}{

\if{html}{\figure{Logo.png}{options: width=120}}
}

\examples{
# Example 1: Subset the first three trackways of MountTom dataset.
subset_data <- subset_track(MountTom, tracks = c(1:3))

# Example 2:  Subset the trackways at indices 5, 7, and 10.
subset_data <- subset_track(MountTom, tracks = c(5, 7, 10))

}
\seealso{
\code{\link{tps_to_track}}
}
\author{
Humberto G. Ferrón

humberto.ferron@uv.es

Macroevolution and Functional Morphology Research Group (www.macrofun.es)

Cavanilles Institute of Biodiversity and Evolutionary Biology

Calle Catedrático José Beltrán Martínez, nº 2

46980 Paterna - Valencia - Spain

Phone: +34 (9635) 44477
}
