% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_prob.R
\name{combined_prob}
\alias{combined_prob}
\title{Calculate combined probabilities of similarity or intersection metrics of tracks}
\usage{
combined_prob(data, metrics = NULL)
}
\arguments{
\item{data}{A 'track' R object, which is a list consisting of two elements:
\itemize{
\item \strong{Trajectories}: A list of interpolated trajectories, where each trajectory is a series of midpoints between consecutive footprints.
\item \strong{Footprints}: A list of data frames containing footprint coordinates, metadata (e.g., image reference, ID), and a marker indicating whether the footprint is actual or inferred.
}}

\item{metrics}{A list of 'track similarity' and/or 'track intersection' R objects derived from different tests. All tests must be
based on the same number of simulations.}
}
\value{
A list containing:
\item{P_values (model names)}{A matrix of p-values for the combined metrics across all trajectories. Each entry represents
the probability of observing the combined metrics between the corresponding pair of trajectories.}
\item{P_values_combined (model names)}{A numeric value representing the overall probability of observing the combined metrics,
across all pairs of trajectories.}
}
\description{
\code{combined_prob()} calculates the combined probabilities of similarity and intersection metrics
derived from different models. The function uses simulation data to extract p-values, providing insight into
the significance of combined metrics across various similarity assessments.
}
\details{
The \code{combined_prob()} function combines p-values derived from multiple similarity metric tests and intersection tests.
It calculates the combined p-values by assessing the probability of observing the combined metrics across simulated datasets.
This function is particularly useful for comparing multiple models and evaluating their collective performance in terms of p-values.
}
\section{Logo}{

\if{html}{\figure{Logo.png}{options: width=30\%}}
}

\examples{
# Example 1: Simulating tracks using the "Directed" model and comparing similarity metrics using DTW, Frechet, and intersection tests.
s1 <- simulate_track(PaluxyRiver, nsim = 1000, model = "Directed")
DTW1 <- simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s1, superposition = "None")
Frechet1 <- simil_Frechet_metric(PaluxyRiver, test = TRUE, sim = s1, superposition = "None")
int1 <- track_instersection(PaluxyRiver, test = TRUE, sim = s1, origin.permutation = "None")
combined_prob(PaluxyRiver, metrics = list(DTW1, Frechet1, int1))

# Example 2: Simulating simulates tracks using the "Constrained" model and comparing similarity metrics using DTW, Frechet, and intersection tests.
s2 <- simulate_track(PaluxyRiver, nsim = 1000, model = "Constrained")
DTW2 <- simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s2, superposition = "None")
Frechet2 <- simil_Frechet_metric(PaluxyRiver, test = TRUE, sim = s2, superposition = "None")
int2 <- track_instersection(PaluxyRiver, test = TRUE, sim = s2, origin.permutation = "Min.Box")
combined_prob(PaluxyRiver, metrics = list(DTW2, Frechet2, int2))

# Example 3: Simulating tracks using the "Unconstrained" model and comparing similarity metrics using DTW, Frechet, and intersection tests.
s3 <- simulate_track(PaluxyRiver, nsim = 1000, model = "Unconstrained")
DTW3 <- simil_DTW_metric(PaluxyRiver, test = TRUE, sim = s3, superposition = "None")
Frechet3 <- simil_Frechet_metric(PaluxyRiver, test = TRUE, sim = s3, superposition = "None")
int3 <- track_instersection(PaluxyRiver, test = TRUE, sim = s3, origin.permutation = "Conv.Hull")
combined_prob(PaluxyRiver, metrics = list(DTW3, Frechet3, int3))

# Example 4: Simulating tracks in the MountTom dataset and comparing similarity metrics using the "Centroid" superposition method.
sbMountTom <- subset_track(MountTom, tracks = c(1, 2, 3, 4, 7, 8, 9, 13, 15, 16, 18))
s4 <- simulate_track(sbMountTom, nsim = 1000)
DTW4 <- simil_DTW_metric(sbMountTom, test = TRUE, sim = s4, superposition = "Centroid")
Frechet4 <- simil_Frechet_metric(sbMountTom, test = TRUE, sim = s4, superposition = "Centroid")
int4 <- track_instersection(sbMountTom, test = TRUE, sim = s4, origin.permutation = "Min.Box")
combined_prob(sbMountTom, metrics = list(DTW4, Frechet4, int4))

# Example 5: Simulating tracks in the MountTom dataset and comparing similarity metrics using the "Origin" superposition method and custom coordinates.
sbMountTom <- subset_track(MountTom, tracks = c(1, 2, 3, 4, 7, 8, 9, 13, 15, 16, 18))
s5 <- simulate_track(sbMountTom, nsim = 1000)
DTW5 <- simil_DTW_metric(sbMountTom, test = TRUE, sim = s5, superposition = "Origin")
Frechet5 <- simil_Frechet_metric(sbMountTom, test = TRUE, sim = s5, superposition = "Origin")
area_origin <- matrix(c(50, 5, 10, 5, 10, 20, 50, 20), ncol = 2, byrow = TRUE)
int5 <- track_instersection(sbMountTom, test = TRUE, sim = s5, origin.permutation = "Custom", custom.coord = area_origin)
combined_prob(sbMountTom, metrics = list(DTW5, Frechet5, int5))

}
\seealso{
\code{\link{tps_to_track}}, \code{\link{simulate_track}}, \code{\link{track_intersection}}, \code{\link{simil_DTW_metric}}, \code{\link{simil_Frechet_metric}}
}
\author{
Humberto G. Ferrón

humberto.ferron@uv.es

Macroevolution and Functional Morphology Research Group (www.macrofun.es)

Cavanilles Institute of Biodiversity and Evolutionary Biology

Calle Catedrático José Beltrán Martínez, nº 2

46980 Paterna - Valencia - Spain

Phone: +34 (9635) 44477
}
