% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tps_to_track.R
\name{tps_to_track}
\alias{tps_to_track}
\title{Transform a *.tps file into a 'track' R object}
\usage{
tps_to_track(file, scale = NULL, missing = FALSE, NAs = NULL, R.L.side = NULL)
}
\arguments{
\item{file}{A *.tps file containing \link{x,y} coordinates of footprints in tracks.}

\item{scale}{A numeric value specifying the scale in meters per pixel.}

\item{missing}{A logical value indicating whether there are missing footprints in any track to be interpolated: \code{TRUE}, or \code{FALSE} (the default).}

\item{NAs}{A matrix with two columns indicating which missing footprints will be interpolated.
The first column gives the number of the track containing missing footprints, and the second column gives the number of the footprint that is missing within this track.
The number of rows is equal to the total number of missing footprints in the sample.}

\item{R.L.side}{A character vector specifying the side of the first footprint of each track.
Only needed if \code{missing} is set to \code{TRUE}.
The length of the vector must be equal to the total number of tracks in the sample.
\itemize{
\item \code{"L"}: first footprint corresponds to the left foot.
\item \code{"R"}: first footprint corresponds to the right foot.
}}
}
\value{
A 'track' R object, which is a list consisting of two elements:
\itemize{
\item \strong{Trajectories}: A list of interpolated trajectories, where each trajectory is a series of midpoints between consecutive footprints.
\item \strong{Footprints}: A list of data frames containing footprint coordinates, metadata (e.g., image reference, ID), and a marker indicating whether the footprint is actual or inferred.
}
}
\description{
\code{tps_to_track()} reads a *.tps file containing footprint coordinates of one or several tracks and transforms it into a 'track' R object.
}
\details{
It is highly recommended that the *.tps file is built using the TPS software (Rohlf 2008, 2009).
Tracks with a different number of footprints (i.e., landmarks) are allowed.
This function transforms the coordinates of the footprints of each track into a set of trajectory coordinates.
Each point of the trajectory is calculated as: \deqn{Point_i[x,y]= (Footprint_i[x,y] + Footprint_{i+1}[x,y])/2}

The number of points of the resulting trajectory is \eqn{n_{footprints} - 1}.

If \code{missing} is set to \code{TRUE}, missing footprints can be interpolated.
This interpolation is based on adjacent footprints and the provided side information.
}
\section{Logo}{

\if{html}{\figure{Logo.png}{options: width=30\%}}
}

\examples{
# Example 1: Tracks without missing footprints. Based on the Paluxy River
# dinosaur chase sequence (Farlow et al., 2011).

# Create a temporary file to store the TPS data. This file will be
# automatically deleted when the R session ends.
tps.file <- tempfile(fileext = ".data")

# Use the cat function to write TPS data directly into the temporary file.
# This TPS data includes footprint coordinates for two different tracks,
# along with associated metadata.
cat(
  "LM=30",
  "211.00000 146.00000", "137.00000 327.00000", "246.00000 447.00000",
  "201.00000 609.00000", "291.00000 704.00000", "246.00000 851.00000",
  "361.00000 959.00000", "327.00000 1129.00000", "438.00000 1226.00000",
  "402.00000 1383.00000", "518.00000 1482.00000", "440.00000 1584.00000",
  "528.00000 1696.00000", "469.00000 1852.00000", "563.00000 1967.00000",
  "497.00000 2112.00000", "597.00000 2211.00000", "550.00000 2365.00000",
  "647.00000 2492.00000", "576.00000 2643.00000", "656.00000 2804.00000",
  "560.00000 2922.00000", "621.00000 3046.00000", "530.00000 3135.00000",
  "588.00000 3291.00000", "496.00000 3430.00000", "566.00000 3570.00000",
  "470.00000 3704.00000", "537.00000 3845.00000", "436.00000 4014.00000",
  "IMAGE=Sauropod.png", "ID=0",
  "LM=25",
  "43.00000 317.00000", "125.00000 463.00000", "102.00000 587.00000",
  "162.00000 745.00000", "170.00000 897.00000", "249.00000 1044.00000",
  "253.00000 1206.00000", "344.00000 1349.00000", "331.00000 1494.00000",
  "386.00000 1612.00000", "358.00000 1753.00000", "413.00000 1886.00000",
  "391.00000 2027.00000", "470.00000 2139.00000", "486.00000 2341.00000",
  "491.00000 2538.00000", "550.00000 2657.00000", "485.00000 2794.00000",
  "516.00000 2921.00000", "460.00000 3065.00000", "506.00000 3231.00000",
  "507.00000 3388.00000", "566.00000 3574.00000", "539.00000 3752.00000",
  "495.00000 3945.00000", "IMAGE=Theropod.png", "ID=1",
  file = tps.file, sep = "\n"
)

# Call the tps_to_track function to convert the TPS data in the file
# into a track object. The 'scale' argument sets the scaling factor
# for the coordinates, and 'missing=FALSE' indicates that no landmarks
# are missing in the dataset.
tps_to_track(tps.file, scale = 0.004341493, missing = FALSE, NAs = NULL)


# Example 2: Tracks with missing footprints. Based on dinosaur tracks from
 # the Mount Tom (Ostrom, 1972).

# Create a temporary file to store the TPS data. This file will be
# automatically deleted when the R session ends.
tps.file <- tempfile(fileext = ".data")

# Define a vector of lines representing the TPS data. Each line in this vector
# represents either the landmark coordinates for a specific image or metadata
# about the image (such as the image filename and ID).
data_lines <- c(
  "LM=10", "9328.00000 3441.00000", "9113.00000 3587.00000", "8982.00000 3782.00000",
  "8794.00000 3907.00000", "8637.00000 4080.00000", "8480.00000 4230.00000",
  "8361.00000 4427.00000", "8157.00000 4503.00000", "8048.00000 4659.00000",
  "7834.00000 4804.00000", "IMAGE=Track 01.png", "ID=0", "LM=10",
  "8957.00000 3299.00000", "8809.00000 3459.00000", "8674.00000 3603.00000",
  "8490.00000 3782.00000", "8339.00000 3945.00000", "8192.00000 4170.00000",
  "8051.00000 4415.00000", "7925.00000 4578.00000", "7753.00000 4754.00000",
  "7640.00000 4976.00000", "IMAGE=Track 02.png", "ID=1", "LM=6",
  "8731.00000 3412.00000", "8543.00000 3638.00000", "8355.00000 3804.00000",
  "8129.00000 4070.00000", "7944.00000 4290.00000", "7778.00000 4459.00000",
  "IMAGE=Track 03.png", "ID=2", "LM=6", "8148.00000 3518.00000",
  "7982.00000 3741.00000", "7822.00000 3967.00000", "7715.00000 4183.00000",
  "7549.00000 4355.00000", "7374.00000 4547.00000", "IMAGE=Track 04.png",
  "ID=3", "LM=3", "7383.00000 3161.00000", "7254.00000 3365.00000",
  "7073.00000 3578.00000", "IMAGE=Track 05.png", "ID=4",
  "LM=4", "7082.00000 3192.00000", "6932.00000 3374.00000", "6743.00000 3547.00000",
  "6577.00000 3741.00000", "IMAGE=Track 06.png", "ID=5",
  "LM=7", "7113.00000 3293.00000", "7032.00000 3496.00000", "6922.00000 3801.00000",
  "6790.00000 3970.00000", "6665.00000 4186.00000", "6508.00000 4343.00000",
  "6414.00000 4525.00000", "IMAGE=Track 07.png", "ID=6",
  "LM=7", "6653.00000 3324.00000", "6446.00000 3525.00000", "6320.00000 3779.00000",
  "6160.00000 4007.00000", "6082.00000 4243.00000", "5903.00000 4418.00000",
  "5762.00000 4719.00000", "IMAGE=Track 08.png", "ID=7",
  "LM=6", "6461.00000 3973.00000", "6775.00000 3775.00000", "7069.00000 3666.00000",
  "7383.00000 3543.00000", "7649.00000 3446.00000", "7878.00000 3374.00000",
  "IMAGE=Track 09.png", "ID=8", "LM=3", "5405.00000 3785.00000",
  "5543.00000 3534.00000", "5671.00000 3337.00000", "IMAGE=Track 10.png",
  "ID=9", "LM=3", "5182.00000 3860.00000", "5305.00000 3616.00000",
  "5411.00000 3362.00000", "IMAGE=Track 11.png", "ID=10",
  "LM=3", "5342.00000 3186.00000", "5276.00000 3427.00000", "5145.00000 3694.00000",
  "IMAGE=Track 12.png", "ID=11", "LM=6", "3339.00000 2221.00000",
  "3270.00000 2155.00000", "3183.00000 2062.00000", "3113.00000 2023.00000",
  "3066.00000 1913.00000", "3019.00000 1832.00000", "IMAGE=Track 13.png",
  "ID=12", "LM=4", "4535.00000 1652.00000", "4430.00000 1911.00000",
  "4415.00000 2113.00000", "4347.00000 2333.00000", "IMAGE=Track 14.png",
  "ID=13", "LM=6", "3962.00000 1950.00000", "3857.00000 2132.00000",
  "3765.00000 2387.00000", "3660.00000 2637.00000", "3512.00000 2845.00000",
  "3392.00000 3089.00000", "IMAGE=Track 15.png", "ID=14",
  "LM=9", "3711.00000 1603.00000", "3638.00000 1796.00000", "3527.00000 1971.00000",
  "3407.00000 2164.00000", "3289.00000 2382.00000", "3146.00000 2588.00000",
  "2992.00000 2742.00000", "2840.00000 2958.00000", "2664.00000 3061.00000",
  "IMAGE=Track 16.png", "ID=15", "LM=3", "2371.00000 2826.00000",
  "2332.00000 2770.00000", "2279.00000 2721.00000", "IMAGE=Track 17.png",
  "ID=16", "LM=6", "3069.00000 1687.00000", "2934.00000 1971.00000",
  "2895.00000 2230.00000", "2750.00000 2475.00000", "2664.00000 2755.00000",
  "2578.00000 2926.00000", "IMAGE=Track 18.png", "ID=17",
  "LM=3", "2214.00000 1098.00000", "2067.00000 1462.00000", "1962.00000 1725.00000",
  "IMAGE=Track 19.png", "ID=18", "LM=4", "1630.00000 1753.00000",
  "1570.00000 2027.00000", "1495.00000 2282.00000", "1416.00000 2500.00000",
  "IMAGE=Track 20.png", "ID=19", "LM=4", "1000.00000 1410.00000",
  "817.00000 1536.00000", "601.00000 1736.00000", "437.00000 1920.00000",
  "IMAGE=Track 21.png", "ID=20", "LM=4", "1265.00000 1335.00000",
  "1125.00000 1446.00000", "1089.00000 1661.00000", "975.00000 1866.00000",
  "IMAGE=Track 22.png", "ID=21", "LM=3", "1928.00000 2874.00000",
  "2392.00000 2837.00000", "2726.00000 2736.00000", "IMAGE=Track 23.png",
  "ID=22"
)

# Write the data lines to the temporary TPS file. This TPS file now contains all
# the data necessary to define the landmark coordinates and associated metadata
# for each track.
writeLines(data_lines, con = tps.file)

# Define a matrix representing the footprints that are missing from the dataset.
# In this example, the matrix 'NAs' specifies that footprint 7 is missing in track 3.
NAs <- matrix(c(7, 3), nrow = 1, ncol = 2)

# Call the tps_to_track function, which will convert the TPS data in the file
# to a track object. The 'scale' argument sets the scaling factor for the coordinates,
# 'missing' specifies whether missing footprints should be handled, 'NAs' provides
# the missing footprints matrix, and 'R.L.side' specifies which side (Right or Left)
# is the first footprint of each track.
tps_to_track(tps.file, scale = 0.004411765, missing = TRUE, NAs = NAs,
             R.L.side = c("R", "L", "L", "L", "R", "L", "R", "R", "L", "L", "L",
                          "L", "L", "R", "R", "L", "R", "R", "L", "R", "R",
                          "R", "R"))

}
\references{
Farlow, J. O., O’Brien, M., Kuban, G. J., Dattilo, B. F., Bates, K. T., Falkingham, P. L., & Piñuela, L. (2012). Dinosaur Tracksites of the Paluxy River Valley (Glen Rose Formation, Lower Cretaceous), Dinosaur Valley State Park, Somervell County, Texas. In Proceedings of the V International Symposium about Dinosaur Palaeontology and their Environment (pp. 41-69). Burgos: Salas de los Infantes.

Ostrom, J. H. (1972). Were some dinosaurs gregarious?. Palaeogeography, Palaeoclimatology, Palaeoecology, 11(4), 287-301.

Rohlf, F. J. 2008. TPSUTIL. Version 1.40. Department of Ecology and Evolution, State University of New York.
Available at \url{https://sbmorphometrics.org/}.

Rohlf, F. J. 2009. tpsDig. Version 2.14. Department of Ecology and Evolution, State University of New York.
Available at \url{https://sbmorphometrics.org/}.
}
\author{
Humberto G. Ferrón

humberto.ferron@uv.es

Macroevolution and Functional Morphology Research Group (www.macrofun.es)

Cavanilles Institute of Biodiversity and Evolutionary Biology

Calle Catedrático José Beltrán Martínez, nº 2

46980 Paterna - Valencia - Spain

Phone: +34 (9635) 44477
}
